Move transformations under Transformation namespace, cleanup
