+++ src/Behat/Behat/Transformation/Call/RuntimeTransformation.php
@@ -16,6 +16,8 @@ use Behat\Testwork\Call\RuntimeCallee;
 /**
  * Transformation that is created and executed in the runtime.
  *
+ * @deprecated Will be removed in 4.0. Use specific transformations instead
+ *
  * @author Konstantin Kudryashov <ever.zet@gmail.com>
  */
 final class RuntimeTransformation extends RuntimeCallee implements Transformation
